move unit tests in hopper folder to other places
